Consolidate realtime tests
